requirements.txt
keras
numpy
pillow
Flask
h5py
matplotlib
tensorflow==1.6.0